Fix TS errors for store/ and Demo user component
