+++ frontend/src/components/auth/Demo/index.tsx
@@ -1,9 +1,10 @@
 import React from "react";
 import { useDispatch } from "react-redux";
 import { login } from "../../../store/session";
+import { AppDispatch } from "../../../store";
 
 function Demo() {
-  const dispatch = useDispatch();
+  const dispatch: AppDispatch = useDispatch();
   const demoLogin = async (event: React.SyntheticEvent) => {
     event.preventDefault();
     const username = "jon";
